+++ R/dst_nbinom.R
@@ -42,8 +42,7 @@ dst_nbinom <- function(size, prob) {
     variance = (1 - prob) * size / prob^2,
     skewness = (2 - prob) / sqrt((1 - prob) * size),
     kurtosis_exc = 6 / size + prob^2 / ((1 - prob) * size),
-    range = c(0, Inf),
     .name = "Negative Binomial",
-    .vtype = "discrete"
+    .support = discrete(natural0())
   )
 }