+++ tests/testthat/test-property-kurtosis.R
@@ -37,7 +37,7 @@ test_that("Kurtosis algorithm matches known vals", {
           c("Hypergeometric", "Bernoulli", "Binomial", "Finite")
         ) {
           # Finite support.
-          expect_error(algorithm_kurtosis(d))
+          expect_equal(suppressMessages(algorithm_kurtosis(d)), supposed_kurt, tolerance = 1e-6)
           r <- range(d)
           x <- seq(r[1], r[2], by = 1L)
           if (pretty_name(d) == "Finite") {
@@ -54,7 +54,7 @@ test_that("Kurtosis algorithm matches known vals", {
           c("Negative Binomial", "Poisson", "Geometric")
         ) {
           # Infinite support.
-          expect_error(algorithm_kurtosis(d))
+          expect_equal(suppressMessages(algorithm_kurtosis(d)), supposed_kurt, tolerance = 1e-6)
           mu <- mean(d)
           sigma <- stdev(d)
           to_add <- Inf